Switch to pre-analysis rather than on-the-fly
This is a key change from the way Clojure is implemented, but it's
necessary for the analysis jank will need to do. It means duplication of
some things, such as the known vars, between the runtime and the
analysis contexts, but that's separation is honestly better for my
sanity anyway.