Track vars and constants per frame
Previously, the whole analysis context was used to track lifted vars and
constants. This results on unoptimized codegen, since the context
doesn't distinguish which nested fn needed each var/constant. Instead,
this is done on a per-frame level now, which fits nicely alongside the
new capture tracking there. The result is not functionally different,
but the generated code is much cleaner and will be more efficient.