Add support for variadic arities
They behave just like Clojure, as far as my testing shows. Arg packing
happens automagically.

This commit also introduces call analysis, which traces the call source
to a fn definition, if possible, and then performs arity matching
statically. This is good not only for catching errors, but also more
efficiently generating param packing code for variadic calls. When this
can't happen, such as when the callable is a parameter, this resolution
needs to happen at run-time, using a new set of `dynamic_call` fns.