Eval each form during analysis
The previous approach of codegen-everything-and-JIT was incorrect. We
need to eval as we go, so defs change, fns are called, etc. Clojure
includes an interpreter, of sorts, combined with the JIT compiler, so we
add a similar thing here.

I got a bit carried away here and also wrote a ton of new tests in jank,
rather than C++, removing the old C++ tests. Combined with huge
optimizations to the jank test runner, the new tests run significantly
faster and have _much_ better coverage.

This also fixes a bunch of stuff around variadic fn calls, packing, let
shadowing, and so on which I found while doing all of this.