Detect vector literals during analysis and optimize codegen
Vector literals here meaning literal vectors which are composed of only
primitive literals. There's a TODO remaining here to optimize the
codegen so that the literals used within the vector are lifted first and
then used to make the vector. Right now, those literals are duplicated,
resulting in extra allocations.