Build clang-tidy with LLVM; fix all linting errors
We can't rely on the system's Clang-tidy, since it may be a different
version from the LLVM we're using. This manifests as it not being able
to read the PCH jank uses for compilation, which means it can't lint
anything. Since ROOT's LLVM doesn't have the extra Clang tools, we need
to clone the upstream LLVM and then link them into ROOT's LLVM. The rest
works well, though.