Allow jank to build a PCH after being installed
PCHs, unfortunately, are not distributable, since they contain absolute
paths for header files. Without a PCH, jank's startup time would be
around 3-4 seconds. With one, right now, it's more like 300ms, so I see
it as essential.

One issue which may arise here is that the user running jank for the
first time will need perms to write to where jank is installed. If this
becomes an issue, we can add support for saving this someplace local,
like ~/.jank as well.