Add assoc read/write interface
This includes implementing `get` and `assoc`. Each of those lead to
implementing `not`, `nnext`, and `second`.

Also, a codegen bug was found for recurs nested within lets/dos. An
explicit continue is generated now. The let/do will generate a return
after the continue, since it's dumb, but that's more an aesthetic issue
for now.