Add typed overloads for math fns
This plays on the previous work to add more type info to generated
constants and to use `auto` more pervasively. When the type of number is
known, we can avoid some polymorphic aspects of Clojure's numeric work.
If all types are known, we can skip right to the actual arithmetic with
no dynamic dispatch at all.

This is the next step leading up to unboxed math support.