Add unboxed math overloads
This allows, currently, nested expressions like `(/ 1.0 (+ n 0.5))` to
avoid boxing for the `(+ n 0.5)` and just return a `native_real` to be
used in the division. From there, the division will be boxed depending
on if it's within another math operation.

Right now, the places where boxing is optional are:

1. Within math operations (`+`, `-`, `*`, `/`, `min`, `max`, `abs`, `<`,
   `<=`, `>`, `>=`, `sqrt`, `pow`, `int`)
2. Within if conditions, which plays nicely with fns like `<`

However, two key things remain to be done:

1. Elide boxing in a `let` when the value doesn't need it in its later
   uses (or provide both, if both are needed)
2. Elide boxing from lifted constants when possible (or provide both, if
   both are needed)