Add next-in-place and fresh-seq; optimize reduce
The new `fresh_seq` fn on the `seqable` behavior is expected to return a
new seq not referenced by anyone else. This allows confident usage of
`next_in_place` on it, so only one allocation is needed for traversal,
regardless of the length of the seq. Both of these have been added into
clojure.core vars so that jank code can take advantage of the benefits.

Reduce, rather than using the new vars, has been rewritten to be native,
for the sake of even better performance.