Start tracking box needs in sem analysis
The tracking is minimal, so far, and is not applied very well to
codegen. Right now, anything that can be unboxed is, but I'm not
tracking when it also needs boxing, and I'm compensating with a lot of
extra `make_box` calls, so this ends up being slower than before this
change. However, it's a stepping stone.

Next up is tracking boxing needs in each sem analysis function, as well
as tracking usages of locals in those functions and whether or not those
usages require boxes. This will give a more complete picture for
codegen, which should allow us to eliminate the extra `make_box` calls.