Improve box tracking and box-related codegen
This also adds a test suite for box-related semantic analysis, since I
kept fixing on thing and breaking another. In order to be able to sanely
check the AST-related data, I've added a new system to convert compiler
data (like AST nodes) to runtime data. The tests can then just use
runtime functions and normal Clojure data operations. In the future,
these could be written in jank, rather than C++.

I've also fixed the codegen output in main, by using the new
`analyze_string` function on the RT context, combined with the
expression wrapping in `evaluate`. It now works exactly as it should,
since it evaluates as it goes. This is a bigger win than it may seems,
since it's a step toward AOT compilation.