Add persistent_hash_map
Previously, jank only supported array maps, which had a max size of 8
k/v pairs. Now those get promoted to a hash map. Each of these has a
distinct object type and sequence object type, but they share a good
chunk of the same implementation through base types.

Since there are multiple types of map now, meta maps had to be
type-erased; we can't know which kind of map it will be.

Also, a nasty, bug was lingering in the native_array_map implementation,
to do with incorrect indices. That was found and fixed.