Add command-line parsing and RT profiling
The command-line parsing applies to all existing features and some not
yet implemented. The RT profiling will be exanded with some additional
tooling, going forward, and I have some scripts already done which
aren't quite ready to commit. We'll be able to easily build charts by
just adding `--profile` to the jank invocation and then running a script
on the generated file. The goal is for this to apply to both compiler
and RT internals, as well as user code. We'll see, once there's actually
user code.

Part of this CLI parsing is the addition of a very simple REPL, coming
from `jank repl`. There's no readline, history, completion, coloring, or
server support yet, but it allows for some quick jankin' around.