Fix unboxed analysis when loading modules
There was an issue when loading pre-compiled (.cpp or otherwise)
modules, where we'd try to use analysis information on fns within that
module to determine if we can unbox values. However, since we didn't
analyze the jank code for that module in this session, we don't have
that info. Instead, we should just rely on the meta and the var in that
scenario. When we do have the analysis info, though, we make sure to
verify what we can.