Fix ambiguity in dynamic fn calling
When dynamically calling a function, we need to know three things:

1. Is the function variadic?
2. Is there an ambiguous fixed overload?
3. How many fixed arguments are required before the packed args?

We cannot perform the correct call without all of this information. Since function calls
are on the hottest path there is, we pack all of this into a single byte. Questions
1 and 2 each get a bit and question 3 gets 4 bits to store the fixed arg count.

From there, when we use it, we strip out the bit for question 2 and we switch/case on
the rest. This allows us to do a O(1) jump on the combination of whether it's variadic
and the required fixed args. Finally, we only need the question 2 bit to disambiguate
one branch of each switch.

The ambiguity comes in this case:

```
(defn ambiguous
  ([a] 1)
  ([a & args] args))
(ambiguous :a)
```

When we call `ambiguous` with a single arg, we want it to match the fixed unary arity.
However, given just questions 1 and 3, we will see that we've met the required args
and that the function is variadic and we'll instead dispatch to the variadic arity, with
an empty sequence for `args`.