Add token info to parse results
We're not using meta here, since not every object has meta. We only
include the start/end of each parsed item, though, so a list of numbers
will only have the ( and ) tokens, not tokens for each number. I'm not
sure how we'd handle that if we need it, but I don't need it yet anyway.