Unify jank's exception reporting
This also adds support for specific macroexpansion exception handling.
Stack traces are only shown when an exception is thrown, rather than
showing for every uncaught error. This implicitly distinguishes compiler
errors from runtime errors, but the main effect is that it cleans up the
error output to only show the stack trace when it's needed.