Unify jank's exception reporting  (#309)
* Reapply "Use cpptrace formatter for stripping syms"

This reverts commit 1ed1c83431894ffbdfcfd781eabe3fa5a7586be9.

* Unify jank's exception reporting

This also adds support for specific macroexpansion exception handling.
Stack traces are only shown when an exception is thrown, rather than
showing for every uncaught error. This implicitly distinguishes compiler
errors from runtime errors, but the main effect is that it cleans up the
error output to only show the stack trace when it's needed.

* Add error reporting test for eval