Fix remaining pointer-from-register issues
Ultimatley, this boiled down to making sure that pointer values are
always backed by an alloc. Then, making sure we always load from allocs.
As long as we're consistent about this, everything fits together nicely.

Previously, I was being too stingy with allocs and this lead to some
pointers being in registers and others not. Reconciling this is doable,
but incredibly error prone and not worth it right now.