Fix RTTI linker error
The way we were handling exception type info was not correct, which then
led to more incorrect code being added to make up for that.

It started with using typeid(T).name() as the source for our RTTI
symbol, but that's a user-readable symbol. It's not meant to be the same
as the RTTI symbol, even if you mangle it again. We can tell this by the
Itanium ABI prefix.

The linker errors we saw were for `N4jank7runtime4orefINS0_6objectEEE`,
which on its own raises some flags. Type info symbols should start with
`_ZTI` and type info names should start with `_ZTS`. This is neither,
since we just took the human readable `.name()` output and then mangled
it. That's what was wrong initially.

Now, in order to get JIT linking working with this incorrect symbol, we
added some extra code to `jit_processor.cpp` to register the symbol into
the runtime and to put the correct type info behind it. This allowed
everything to link, but it was a hack. The JIT runtime already has
access to all symbols within the process and we're compiling with
exported symbols and RTTI, so no extra steps should be needed within
`jit_processor.cpp`. Indeed, in this commit, all of that code has been
removed.

There is no actual way to get the `_ZTI` symbol given a `typeinfo`
object. We definitely cannot add `_ZTI`, since this mangling scheme is
only used by the Itanium ABI and will not work on any Microsoft
compiler. Instead, we need to get a `QualType` and then rely on Clang to
portably handle getting us the RTTI symbol for that type. This led to a
new function in the CppInterOp lib called `MangleRTTI`, which takes an
opaque `QualType` and gives us back a string.

Finally, in order to prepare for the upcoming changes which will support
custom exception types, I have added a `type` member to the `catch`
expression. Now we don't need any codegen changes for custom type
support. We just need to analyze the type and plug it into that member.