Explicitly register RTTI on macOS
This is a whole dance which is a supreme pain in the ass. macOS doesn't
export RTTI by default, though Linux does. What this means is two fold,
which each fold being further pain.

1. We need to register RTTI into the JIT
2. We need to generate RTTI into our object files so they can be loaded
   without issue

Neither of these can work in isolation, since jank works both JIT and
AOT.

We tackle the first pain using some some ORC APIs to define symbols
into the current JIT runtime and we rely on JIT compiling some C++ to
get the RTTI addresses, since we don't know the types we're working with
until runtime.

We tackle the second pain by keeping track of all the RTTI a function
uses during codegen and passing that up to higher codegen processors,
all the way up to the module level. At that level, we link in all of
those RTTI modules so that the globals will make it into the generated
object file with definitions, as opposed to an external global. This
keeps our object files standalone.

Now, in some cases, such as debug builds, the JIT runtime already finds
some RTTI symbols, so we fail to add them with redefinition errors.
Careful caching and error consumption allows this to still work.

What a mess.