codegen for "try ... catch ... finally" with native llvm exception handling (#345)
* compile works, test still fails

* compile works, nested try fails

* fixed nested try with inner throw and outer catch; fixed try with empty catch body

* rethrow exception if no catch

* eveything works

* cosmetic fixes

* add a failing test case

* resolve merge conflict

* remove legacy jank_try

* cosmetic fixes

* cosmetic fixes

* fix finally body position to statement

* fix infinite loop

* fix nested try catch

* address comments; add unicode test file;

* address comments

* add more test cases

* update test case

* Upgrade to LLVM 22

* Access raw LLVM context the upstream way

* Build from latest LLVM 22 snapshot

* Use latest libc++ and libunwind on mac

* Specify exact sqrt overload

* Repoint cppinterop

* Bump flake to LLVM 22

* Use correct Clang cache

* Bump CI tool LLVM version

* Get everything compiling

* Cleanup

* Fix linting issues

* Install different LLVM libs on macOS

* Fix clang-tidy issues

* Ensure finally bodies are statement position

* Address some feedback

* clean up and update comments

* only add a personality function when none exists

* restruct the IR gen to be more intuitive

also, add extensive comments explaining the whole process

* remove unused headers and sort headers

* check if we have a terminator directly

* revert return nullptr

* Fix RTTI linker error

The way we were handling exception type info was not correct, which then
led to more incorrect code being added to make up for that.

It started with using typeid(T).name() as the source for our RTTI
symbol, but that's a user-readable symbol. It's not meant to be the same
as the RTTI symbol, even if you mangle it again. We can tell this by the
Itanium ABI prefix.

The linker errors we saw were for `N4jank7runtime4orefINS0_6objectEEE`,
which on its own raises some flags. Type info symbols should start with
`_ZTI` and type info names should start with `_ZTS`. This is neither,
since we just took the human readable `.name()` output and then mangled
it. That's what was wrong initially.

Now, in order to get JIT linking working with this incorrect symbol, we
added some extra code to `jit_processor.cpp` to register the symbol into
the runtime and to put the correct type info behind it. This allowed
everything to link, but it was a hack. The JIT runtime already has
access to all symbols within the process and we're compiling with
exported symbols and RTTI, so no extra steps should be needed within
`jit_processor.cpp`. Indeed, in this commit, all of that code has been
removed.

There is no actual way to get the `_ZTI` symbol given a `typeinfo`
object. We definitely cannot add `_ZTI`, since this mangling scheme is
only used by the Itanium ABI and will not work on any Microsoft
compiler. Instead, we need to get a `QualType` and then rely on Clang to
portably handle getting us the RTTI symbol for that type. This led to a
new function in the CppInterOp lib called `MangleRTTI`, which takes an
opaque `QualType` and gives us back a string.

Finally, in order to prepare for the upcoming changes which will support
custom exception types, I have added a `type` member to the `catch`
expression. Now we don't need any codegen changes for custom type
support. We just need to analyze the type and plug it into that member.

* Remove dead code

* Codegen cleanup

* Try working around macOS _ prefix

* Add missing c_str

* Explicitly register RTTI on macOS

This is a whole dance which is a supreme pain in the ass. macOS doesn't
export RTTI by default, though Linux does. What this means is two fold,
which each fold being further pain.

1. We need to register RTTI into the JIT
2. We need to generate RTTI into our object files so they can be loaded
   without issue

Neither of these can work in isolation, since jank works both JIT and
AOT.

We tackle the first pain using some some ORC APIs to define symbols
into the current JIT runtime and we rely on JIT compiling some C++ to
get the RTTI addresses, since we don't know the types we're working with
until runtime.

We tackle the second pain by keeping track of all the RTTI a function
uses during codegen and passing that up to higher codegen processors,
all the way up to the module level. At that level, we link in all of
those RTTI modules so that the globals will make it into the generated
object file with definitions, as opposed to an external global. This
keeps our object files standalone.

Now, in some cases, such as debug builds, the JIT runtime already finds
some RTTI symbols, so we fail to add them with redefinition errors.
Careful caching and error consumption allows this to still work.

What a mess.

* Remove some dead code

* Fix clang-tidy issue

* Address more clang-tidy issues

---------

Co-authored-by: jeaye <contact@jeaye.com>