fix(module-loader): Remove `loaded-libs` check in loader::load()
We introduced checking whether a module is loaded already i.e. 
exists in `clojure.core/*loaded-libs*`, in
`jank::runtime::module::loader::load()` to ensure clojure.core when
loaded from the functions in main.cpp, is only loaded once.

Clojure only interacts with `*loaded-libs` from its `clojure.core` ns.
jank needs to interact with it from our cpp source because there we
directly rely on `jank::runtime::context::load_module` instead of going
through `(clojure.core/load-lib ...)`.

Now that within the main, when loading `clojure.core`, we also mark it
as loaded in `*loaded-libs*`, we don't need this check. In the
hindsight, this approach is a better one than modifying the
`loader::load()` logic.

Reverting this logic, does have a bit of performance impact. The reason
being not being short-circuited when the module already is loaded within
the `loader::load()` function.

This is how Clojure does it right now. We'll see later, if this should be
optimized.